server/pyproject.toml
[tool.poetry]
name = "auraed"
version = "0.1.0"
description = ""
authors = ["CaffeineDuck <hello@samrid.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
tortoise-orm = {extras = ["asycnpg"], version = "^0.19.1"}
fastapi = {extras = ["all"], version = "^0.78.0"}
aerich = "^0.6.3"
pydantic = "^1.9.1"
asyncpg = "^0.25.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
pre-commit = "^2.19.0"
aerich = "^0.6.3"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.pylint.main]
typeCheckingMode = "basic"

[tool.aerich]
tortoise_orm = "tortoise_config.tortoise_config"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
